requirements.txt
streamlit==1.36.0
langchain==0.2.11
langchain-groq==0.1.6
python-dotenv==1.0.1